requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.28.70
botocore==1.31.70
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.4
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==4.2.5
et-xmlfile==1.1.0
executing==2.0.0
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
html5==0.0.9
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipykernel==6.25.2
ipython==8.16.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.4.0
jupyter_core==5.4.0
lxml==4.9.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
plotly==5.17.0
pprintpp==0.4.0
prompt-toolkit==3.0.39
Protego==0.3.0
psutil==5.9.5
psycopg2==2.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
pyodbc==5.0.1
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
Scrapy==2.11.0
scrapy-selenium==0.0.7
selenium==3.141.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
tldextract==5.1.1
tornado==6.3.3
traitlets==5.11.2
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.25.0
w3lib==2.1.2
wcwidth==0.2.8
webencodings==0.5.1
wsproto==1.2.0
zope.interface==6.1
